move implementation of WritableGroupStorage out of range.go
